restructure PostFormPopUp code
