+++ src/app/(home)/components/animePostComponent/index.js
@@ -1,0 +1,5 @@
+export { default as AnimeEpisodes } from "./AnimeEpisodes/AnimeEpisodes.tsx"
+export { default as AnimeScore } from "./AnimeScore/AnimeScore.tsx"
+export { default as AnimeSearch } from "./AnimeSearch/AnimeSearch.tsx"
+export { default as AnimeTag } from "./AnimeTag/AnimeTag.tsx"
+export { default as AnimeWatchStatus } from "./AnimeWatchStatus/AnimeWatchStatus.tsx"
