+++ src/app/(home)/components/functions/syncUpdates.ts
@@ -61,7 +61,7 @@ export async function updateStatusOnFriendLists(myUserInfo: UserInfo, postAnimeD
   const updatePromises = friendsDocs.map((friendDoc: any) => {
     if (!friendDoc.empty) {
       const docRef = friendDoc.docs[0].ref
-      const oldInfo = friendDoc.docs[0].data().friend_list.find((info: any) => info.username === myUserInfo.username)
+      const oldInfo = friendDoc.docs[0].data().friend_list?.find((info: any) => info.username === myUserInfo.username)
 
       if (!oldInfo) return
       const batch = writeBatch(db)
