+++ src/components/utils/postingUtils.ts
@@ -1,5 +1,7 @@
+import { updateStatusOnFriendLists } from "@/app/(home)/components/functions/syncUpdates"
 import { db } from "@/firebase/firebase-app"
-import { doc, getDoc } from "firebase/firestore"
+import { addDoc, collection, doc, getDoc, serverTimestamp, setDoc } from "firebase/firestore"
+import getProductionBaseUrl from "./getProductionBaseURL"
 
 export function getDateNow() {
   const now = new Date().toISOString().slice(0, -5) + "+00:00"
@@ -84,4 +86,119 @@ export async function adjustAnimeListArray(username: any, animeData: any) {
   }
   // if anime data is in the list
   return newDataArray
+}
+
+export async function handleSubmitForm(
+  e: React.FormEvent<HTMLFormElement>,
+  animeStatus: any,
+  animeSearch: any,
+  animeEpisodes: any,
+  animeTag: any,
+  animeScore: any,
+  inputRef: any,
+  mediaUrl: any,
+  setLoadPosting: any,
+  uploadMedia: any,
+  username: any,
+  avatarUrl: any,
+  mediaType: any,
+  malAuthCode: any,
+  myUserInfo: any,
+): Promise<void> {
+  const statusData = (animeStatus?.current as any).getAnimeStatus()
+  const searchData = (animeSearch?.current as any).getAnimeName()
+  const episodesData = (animeEpisodes?.current as any).getAnimeEpisodes()
+  const totalEps = (animeEpisodes?.current as any).getAnimeTotal()
+  const tagData = (animeTag?.current as any).getAnimeTag()
+  const scoreData = (animeScore?.current as any).getAnimeScore()
+
+  e.preventDefault()
+  if (!inputRef.current?.value && mediaUrl.length === 0) {
+    alert("Please write something or upload media 😭")
+    return
+  }
+  setLoadPosting(true)
+  const postAnimeData = handleAnimeInformationPosting(
+    statusData,
+    searchData,
+    episodesData,
+    tagData,
+    totalEps,
+    scoreData,
+  )
+
+  const downloadMediaUrl = await uploadMedia()
+  //TODO: promise.all post + update MAL watch status
+  const promisePost = [
+    addDoc(collection(db, "posts"), {
+      authorName: username,
+      avatarUrl: avatarUrl,
+      timestamp: serverTimestamp(),
+      content: inputRef.current?.value || "",
+      imageUrl: mediaType === "image" ? downloadMediaUrl : "",
+      videoUrl: mediaType === "video" ? downloadMediaUrl[0] : "",
+      post_anime_data:
+        (animeSearch?.current as any).getAnimeName().animeID === "" ? { tag: postAnimeData.tag } : postAnimeData,
+      comments: 0,
+    }),
+  ] as any
+
+  try {
+    // post have anime data
+    if (!!postAnimeData.anime_id) {
+      // user connect with MAL
+      if (!!malAuthCode) {
+        const promiseUpdateMAL = fetch(getProductionBaseUrl() + "/api/updatestatus/" + postAnimeData.anime_id, {
+          headers: {
+            status: convertWatchStatus(postAnimeData.watching_progress),
+            episode: postAnimeData.episodes_seen,
+            score: (postAnimeData as any).score,
+            auth_code: malAuthCode,
+          } as any,
+        }).then((res) => res.json())
+        promisePost.push(promiseUpdateMAL)
+      } else {
+        // user not connect with MAL
+        const myAnimeListRef = doc(db, "myAnimeList", username)
+        const dateNow = getDateNow()
+        const animeInformation = await fetch(getProductionBaseUrl() + "/api/anime/" + postAnimeData.anime_id).then(
+          (res) => res.json(),
+        )
+        const animeData = {
+          list_status: {
+            is_rewatching: false,
+            num_episodes_watched: postAnimeData.episodes_seen,
+            score: !!(postAnimeData as any).score ? (postAnimeData as any).score : 0,
+            status: convertWatchStatus(postAnimeData.watching_progress),
+            updated_at: dateNow,
+          },
+          node: {
+            id: postAnimeData.anime_id,
+            main_picture: animeInformation.main_picture,
+            title: animeInformation.title,
+          },
+        }
+        // check if anime already in list, if in list, update that anime node
+        const animeDataAfterCheck = await adjustAnimeListArray(username, animeData)
+        const myAnimeListPromise = setDoc(myAnimeListRef, {
+          last_updated: dateNow,
+          animeList: animeDataAfterCheck,
+        })
+        promisePost.push(myAnimeListPromise as any)
+        console.log("user not connect with MAL")
+      }
+      // Update status on friend list
+      promisePost.push(
+        updateStatusOnFriendLists(myUserInfo, {
+          ...postAnimeData,
+          status: convertWatchStatus(postAnimeData.watching_progress),
+        }),
+      )
+    }
+    const result = await Promise.all(promisePost)
+    setLoadPosting(false)
+    location.reload()
+  } catch (err) {
+    console.log(err)
+  }
 }
